Fix UB introduced by c2rust
In the original C code the mode2 variable in bzopen_or_bzdopen is
allocated on the stack. C2rust decided to put it in static memory
instead and transmute it to a mutable reference. This will likely result
in crashes aside from the UB itself as the static memory is immutable
and bzopen_or_bzdopen tries to write to it.