Some changes to Arr2::block_and_quadrant
Using .byte_add() is a bit nicer than cast + .wrapping_add() + cast.
wrapping_add is also not necessary given that we stay within the bounds
of the allocation.

Use <[_]>.fill() instead of write_bytes as the former is safe unlike the
latter. Given that the memory allocation is already zeroed when
allocating it, there is no risk of exposing uninitialized memory.