Use panics in handle_assert_failure too when stdio mode is disabled
When using libbz2-rs-sys as a rust dependency, it is not possible to
define bz_internal_error in a way that GNU ld will accept. The
definition would come before the use in libbz2-rs-sys and GNU ld will
only resolve symbol uses when the defining static library comes after
the static library that needs the symbol.

When using as cdylib bz_internal_error will still be called by the panic
handler in libbz2-rs-sys-cdylib.